Struct paging::arch::table::Entry

pub struct Entry(_);

Methods

impl Entry

fn new(addr: PAddr) -> Self

fn do_huge(&self, offset: usize) -> Option<PhysicalPage>

fn is_unused(&self) -> bool

Returns true if this is an unused entry

fn set_unused(&mut self)

Sets this entry to be unused

fn is_huge(&self) -> bool

Returns true if this page is huge

fn flags(&self) -> EntryFlags

Access the entry's bitflags.

fn get_addr(&self) -> PAddr

Returns the physical address pointed to by this page table entry

fn get_frame(&self) -> Option<PhysicalPage>

Returns the frame in memory pointed to by this page table entry.

fn set(&mut self, frame: PhysicalPage, flags: EntryFlags)