Struct paging::arch::table::Table

#[repr(C)]

pub struct Table<L> where
    L: TableLevel,  { /* fields omitted */ }

A page table

Methods

impl Table<PML4Level>

fn page_table_for(&self, page: VirtualPage) -> Option<&Table<PTLevel>>

fn page_table_mut_for(
&mut self,
page: VirtualPage
) -> Option<&mut Table<PTLevel>>

impl<L: TableLevel> Table<L>

fn zero(&mut self) -> &mut Self

Zeroes out the page table by setting all entries "unused"

fn start_paddr(&self) -> PAddr

Return the start physical address of this Table

fn frame(&self) -> PhysicalPage

Return the PhysicalPage containing this table.

impl<L: Sublevel> Table<L>

fn next_table<I>(&self, i: I) -> Option<&Table<L::Next>> where
L: IndexOf<I>,

Returns the next table, or None if none exists

fn next_table_mut<I>(&self, i: I) -> Option<&mut Table<L::Next>> where
L: IndexOf<I>,
I: Debug,

Mutably borrows the next table.

fn create_next<A>(
&mut self,
i: VirtualPage,
alloc: &mut A
) -> MapResult<&mut Table<L::Next>> where
A: FrameAllocator,

Returns the next table, creating it if it does not exist.