Struct paging::FrameManager

pub struct FrameManager { /* fields omitted */ }

FrameManager verwaltet die Allozierung von Frames. Jedes Bit in bits steht für einen Frame.

Methods

impl FrameManager

fn get() -> &'static mut FrameManager

Gibt eine Referenz auf Seitenverzeichnis-Singleton zurück

fn reserve(&mut self, frm: Frame) -> Result<Frame, FrameError>

Markiert einen Frame als reserviert und gibt ihn im Erfolgsfall zurück. Im Fehlerfall wird der entsprechende FrameError zurückgegeben.

fn allocate(&mut self) -> Result<Frame, FrameError>

Reserviert einen freien Frame und gibt ihn im Erfolgsfall zurück. Im Fehlerfall wird der entsprechende FrameError zurückgegeben.

fn release(&mut self, frm: Frame) -> Result<(), FrameError>

Gibt einen reservierten Frame frei

fn reserve_range(&mut self, r: AddressRange)

Markiert alle Frames eines Adressbereiches als reserviert

Anmerkung

Wenn die Grenzen des Adressbereichs nicht seiten-aligned sind, werden die "Randbereiche" reserviert.